added portkey to chatbot app for testing
