+++ experiments/multiagent_coding/smolagents/chatbot_app/requirements.txt
